refactor(model): move the data models function out
