+++ docs/source/api_doc/model/clazz.rst
@@ -1,0 +1,37 @@
+hbutils.model.clazz
+========================================
+
+.. currentmodule:: hbutils.model.clazz
+
+.. automodule:: hbutils.model.clazz
+
+
+asitems
+---------------------
+
+.. autofunction:: asitems
+
+
+visual
+---------------------
+
+.. autofunction:: visual
+
+
+constructor
+---------------------
+
+.. autofunction:: constructor
+
+
+hasheq
+---------------------
+
+.. autofunction:: hasheq
+
+
+accessor
+---------------------
+
+.. autofunction:: accessor
+
