+++ docs/source/api_doc/model/index.rst
@@ -1,0 +1,12 @@
+hbutils.model
+=====================
+
+.. currentmodule:: hbutils.model
+
+.. automodule:: hbutils.model
+
+.. toctree::
+    :maxdepth: 3
+
+    clazz
+
