+++ docs/source/index.rst
@@ -27,6 +27,7 @@ Overview
     api_doc/config/index
     api_doc/design/index
     api_doc/encoding/index
+    api_doc/model/index
     api_doc/random/index
     api_doc/reflection/index
     api_doc/scale/index
